WasmParser: Add `FileSystem` trait to allow omitting file system helpers
Extracted from https://github.com/swiftwasm/WasmKit/pull/357
