+++ .github/workflows/main.yml
@@ -61,7 +61,7 @@ jobs:
           ./Vendor/checkout-dependency --category component-model
 
           swift --version
-          swift test --sanitize address --traits ComponentModel,WasmDebuggingSupport ${{ matrix.test-args }}
+          swift test --sanitize address --traits FileSystem,ComponentModel,WasmDebuggingSupport ${{ matrix.test-args }}
           swift build --package-path Benchmarks
 
   build-macos:
@@ -71,7 +71,7 @@ jobs:
           # Swift 6.3.2
           - os: macos-26
             xcode: Xcode_26.5
-            test-args: "--sanitize address --traits ComponentModel,WasmDebuggingSupport"
+            test-args: "--sanitize address --traits FileSystem,ComponentModel,WasmDebuggingSupport"
 
     runs-on: ${{ matrix.os }}
     name: "build-macos (${{ matrix.xcode }})"
@@ -132,15 +132,15 @@ jobs:
       matrix:
         include:
           - swift: "swift:6.3-noble"
-            test-args: "--traits ComponentModel,WasmDebuggingSupport --enable-code-coverage"
+            test-args: "--traits FileSystem,ComponentModel,WasmDebuggingSupport --enable-code-coverage"
             build-benchmarks: true
             run-benchmarks: true
             build-dev-dashboard: true
           - swift: "swiftlang/swift:nightly-6.4.x-noble"
-            test-args: "--traits ComponentModel,WasmDebuggingSupport"
+            test-args: "--traits FileSystem,ComponentModel,WasmDebuggingSupport"
             build-benchmarks: true
           - swift: "swiftlang/swift:nightly-main-noble"
-            test-args: "--traits ComponentModel,WasmDebuggingSupport"
+            test-args: "--traits FileSystem,ComponentModel,WasmDebuggingSupport"
             build-benchmarks: true
 
     runs-on: ubuntu-24.04
@@ -234,9 +234,9 @@ jobs:
         run: ./build-exec swift sdk install "${{ matrix.musl-swift-sdk-download }}" --checksum "${{ matrix.musl-swift-sdk-checksum }}"
 
       - name: Build (x86_64-swift-linux-musl)
-        run: ./build-exec swift build --swift-sdk x86_64-swift-linux-musl --traits ComponentModel,WasmDebuggingSupport --explicit-target-dependency-import-check error
+        run: ./build-exec swift build --swift-sdk x86_64-swift-linux-musl --traits FileSystem,ComponentModel,WasmDebuggingSupport --explicit-target-dependency-import-check error
       - name: Build (aarch64-swift-linux-musl)
-        run: ./build-exec swift build --swift-sdk aarch64-swift-linux-musl --traits ComponentModel,WasmDebuggingSupport --explicit-target-dependency-import-check error
+        run: ./build-exec swift build --swift-sdk aarch64-swift-linux-musl --traits FileSystem,ComponentModel,WasmDebuggingSupport --explicit-target-dependency-import-check error
 
   build-android:
     runs-on: ubuntu-24.04
@@ -297,4 +297,4 @@ jobs:
       - name: Install Swift SDK
         run: swift sdk install https://download.swift.org/swift-6.3.2-release/wasm-sdk/swift-6.3.2-RELEASE/swift-6.3.2-RELEASE_wasm.artifactbundle.tar.gz --checksum a61f0584c93283589f8b2f42db05c1f9a182b506c2957271402992655591dd7c
       - name: Build with the Swift SDK
-        run: swift build --swift-sdk "$(swiftc -print-target-info | jq -r '.swiftCompilerTag')_wasm" --traits ComponentModel --explicit-target-dependency-import-check error --product wasmkit-cli
+        run: swift build --swift-sdk "$(swiftc -print-target-info | jq -r '.swiftCompilerTag')_wasm" --traits FileSystem,ComponentModel --explicit-target-dependency-import-check error --product wasmkit-cli
